requirements.txt
aiofiles==0.6.0
certifi==2020.11.8
cffi==1.14.3
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
# Editable Git install with no remote (image-survey==1.0)
-e /home/kaan/Projects/image-survey
multidict==5.0.0
pycparser==2.20
PyJWT==1.7.1
PyYAML==5.3.1
rfc3986==1.4.0
sanic==20.9.1
sanic-jwt==1.5.0
six==1.15.0
sniffio==1.2.0
ujson==4.0.1
uvloop==0.14.0
websockets==8.1
